Avoid rming everything in markbad.sh...

+++ markbad.sh
@@ -4,6 +4,11 @@
 
 name=$(printf "$1" | sed 's@/$@@')
 
+if ! [ -d $outhistdir/$name ]; then
+  echo No hist directory for $name, doing nothing
+  exit 1
+fi
+
 touch $outhistdir/$name/bad
 
 rm -fv $outhadddir/${name}*
